// src/app/api/models/consultation-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ConsultationResponseDto {
  allergies?: Array<string>;
  appointment?: {
};
  appointmentId?: number;
  bloodPressure?: string;
  client?: {
};
  clientId: number;
  consultationNumber: string;
  createdAt: string;
  diagnosis?: string;
  endTime?: string;
  files?: Array<string>;
  followUpDate?: string;
  followUpRequired?: boolean;
  heartRate?: number;
  height?: number;
  id: number;
  invoices?: Array<string>;
  medications?: Array<string>;
  nextAppointment?: string;
  notes?: string;
  recommendations?: string;
  startTime?: string;
  status: 'pending' | 'in_progress' | 'completed' | 'cancelled';
  symptoms?: string;
  temperature?: number;
  treatment?: string;
  updatedAt: string;
  user?: {
};
  userId: number;
  vitals?: {
};
  weight?: number;
}